requirements.txt
flask==3.1.1
matplotlib==3.9.4
numpy==2.0.2
torch==2.8.0
torchvision==0.23.0
pillow==11.3.0
